♻️(encode) spin MediaConvert config off into a separate file
MediaConvert configuration for video encoding was handled in our
main handler in index.js. This was fine as long as we only received
events for video file uploads.

Now that we're preparing to upload subtitle tracks along with video
files, we need to put it into its own function and file so we keep
the main handler clean and simple. We'll also put subtitle track
conversion in its own file as we add it to the lambda.
